requirements.txt
pandas
sqlalchemy
scikit-learn
openai
fastapi
uvicorn
streamlit
faiss-cpu
